pyproject.toml
[project]
name = "crawl-ai"
version = "0.1.0"
description = "AI-Native Intelligence Platform for automated web crawling and analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",

    # Task Queue
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",

    # Crawling
    "httpx>=0.26.0",
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "feedparser>=6.0.0",

    # AI APIs
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "aiofiles>=23.2.0",
    "python-dateutil>=2.8.0",

    # Notifications
    "slack-sdk>=3.26.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
